Consistently use integers instead of shorts for the region id, air code, GIFT material code, and los value for the rt_comb_internal structure.  region ids are already handled almost entirely as integers throughout the code, but the usage of shorts in the comb structure seems to be a carry-over from prior v4 format support that was never changed.  a review of code seems to indicate that the only negative impact should only be rather unsupported behavior if someone actually reverts a v5 db to v4 that has region idents that overflow short.  (this mod is in support of JTCG/ME and addresses sf bug 1532661 "Mismatch In Size of region_id" reported by Tom Browder)
svn:revision:27051
cvs:account:brlcad
cvs:branch:trunk

+++ include/raytrace.h
@@ -920,10 +920,10 @@ struct rt_comb_internal  {
 	char		region_flag;	/* !0 ==> this COMB is a REGION */
 	char		is_fastgen;	/* REGION_NON_FASTGEN/_PLATE/_VOLUME */
 	/* Begin GIFT compatability */
-	short		region_id;
-	short		aircode;
-	short		GIFTmater;
-	short		los;
+	int		region_id;
+	int		aircode;
+	int		GIFTmater;
+	int		los;
 	/* End GIFT compatability */
 	char		rgb_valid;	/* !0 ==> rgb[] has valid color */
 	unsigned char	rgb[3];
